Move code from lib.rs to lookup.rs
